refactor/3-prj-structure-refactor 문제 생성과 챗봇 기능 디렉토리 분리, 파일명 변경, 레거시 코드 삭제
