--- log.txt
@@ -1,35 +1,0 @@
-RAG 체인에 전달할 쿼리: 日本語能力試験（JLPT）のN2レベルに該当する、「科学」に関する読解問題を作成し てください。問題のタイプは「筆者の主張や意図を問う問題」で、文章の長さは「短い文章(約200字)」にしてください。作成する問題の「explanation」フィールドは、必ず韓国語で作成してください。
-JSON to be sent to backend: {
-  "problem_title_parent": "\u6b21\u306e\u6587\u7ae0\u3092\u8aad\u3093\u3067\u3001\u5f8c\u306e\u554f\u3044\u306b\u7b54\u3048\u306a\u3055\u3044\u3002",
-  "problem_title_child": "\u7b46\u8005\u306e\u4e3b\u5f35\u3068\u3057\u3066\u6700\u3082\u9069\u5207\u306a\u3082\u306e\u3092\u30011\u30fb2\u30fb3\u30fb4\u304b\u3089\u4e00\u3064\u9078\u3073\u306a\u3055\u3044\u3002",
-  "problem_content": "\u8fd1\u5e74\u3001\u4eba\u5de5\u77e5\u80fd\uff08AI\uff09\u306e\u767a\u5c55\u306f\u76ee\u899a\u307e\u3057\u3044\u3082\u306e\u304c\u3042\u308b\u3002\u533b\u7642\u8a3a\u65ad\u306e\u88dc\u52a9\u3084\u81ea\u52d5\u904b\u8ee2\u6280\u8853\u306a\u3069\u3001\u69d8\u3005\u306a\u5206\u91ce\u3067AI\u306e\u6d3b\u7528\u304c\u9032\u307f\u3001\u79c1\u305f\u3061\u306e\u751f\u6d3b\u306f\u5927\u304d\u304f\u5909\u308f\u308a\u3064\u3064\u3042\u308b\u3002\u3057\u304b\u3057\u3001AI\u306f\u4e07\u80fd\u3067\u306f\u306a\u3044\u3002AI\u306f\u5927\u91cf\u306e\u30c7\u30fc\u30bf\u306b\u57fa\u3065\u3044\u3066\u5b66\u7fd2\u3057\u3001\u30d1\u30bf\u30fc\u30f3\u8a8d\u8b58\u3084\u4e88\u6e2c\u3092\u884c\u3046\u304c\u3001\u4eba\u9593\u306e\u6301\u3064\u5275\u9020\u6027\u3084\u502b\u7406\u89b3\u3001\u611f\u60c5\u3068\u3044\u3063\u305f\u3082\u306e\u306f\u6301\u3061\u5408\u308f\u305b\u3066\u3044\u306a\u3044\u3002AI\u3092\u904e\u4fe1\u3057\u3001\u4eba\u9593\u304c\u601d\u8003\u3059\u308b\u3053\u3068\u3092\u653e\u68c4\u3059\u308c\u3070\u3001\u793e\u4f1a\u306b\u601d\u308f\u306c\u5f0a\u5bb3\u304c\u751f\u3058\u308b\u53ef\u80fd\u6027\u3082\u3042\u308b\u3002AI\u306f\u3042\u304f\u307e\u3067\u9053\u5177\u3067\u3042\u308a\u3001\u305d\u308c\u3092\u3069\u306e\u3088\u3046\u306b\u6d3b\u7528\u3059\u308b\u304b\u306f\u4eba\u9593\u306e\u8cac\u4efb\u3067\u3042\u308b\u3002AI\u306e\u30e1\u30ea\u30c3\u30c8\u3092\u4eab\u53d7\u3057\u3064\u3064\u3001\u305d\u306e\u9650\u754c\u3092\u7406\u89e3\u3057\u3001\u4eba\u9593\u304c\u4e3b\u4f53\u7684\u306b\u793e\u4f1a\u3092\u7bc9\u3044\u3066\u3044\u304f\u3053\u3068\u304c\u91cd\u8981\u3067\u3042\u308b\u3002",
-  "choices": [
-    {
-      "number": 1,
-      "content": "AI\u306f\u4eba\u9593\u306e\u751f\u6d3b\u3092\u8c4a\u304b\u306b\u3059\u308b\u4e07\u80fd\u306a\u9053\u5177\u3067\u3042\u308b\u3002",
-      "is_correct": false
-    },
-    {
-      "number": 2,
-      "content": "AI\u306f\u4eba\u9593\u306e\u5275\u9020\u6027\u3084\u502b\u7406\u89b3\u3092\u51cc\u99d5\u3059\u308b\u5b58\u5728\u306b\u306a\u308b\u3060\u308d\u3046\u3002",
-      "is_correct": false
-    },
-    {
-      "number": 3,
-      "content": "AI\u306e\u6d3b\u7528\u306f\u793e\u4f1a\u306b\u601d\u308f\u306c\u5f0a\u5bb3\u3092\u3082\u305f\u3089\u3059\u305f\u3081\u3001\u614e\u91cd\u306b\u306a\u308b\u3079\u304d\u3060\u3002",      
-      "is_correct": false
-    },
-    {
-      "number": 4,
-      "content": "AI\u306e\u9650\u754c\u3092\u7406\u89e3\u3057\u3001\u4eba\u9593\u304c\u4e3b\u4f53\u7684\u306bAI\u3092\u6d3b\u7528\u3057\u3066\u3044\u304f\u5fc5\u8981\u304c\u3042\u308b\u3002",
-      "is_correct": true
-    }
-  ],
-  "answer_number": 4,
-  "explanation": "4) AI\uc758 \ud55c\uacc4\ub97c \uc774\ud574\ud558\uace0 \uc778\uac04\uc774 \uc8fc\uccb4\uc801\uc73c\ub85c AI\ub97c \ud65c\uc6a9\ud574 \ub098\uac00\uc57c \ud55c\ub2e4: \ud544\uc790\ub294 AI\uac00 \uc778\uac04\uc758 \uc0b6\uc744 \ubc14\uafc0 \uc720\uc6a9\ud55c \ub3c4\uad6c\uc774\uc9c0\ub9cc, \ucc3d\uc758\uc131\uc774\ub098 \uc724\ub9ac\uad00, \uac10\uc815 \ub4f1 \uc778\uac04\uc801\uc778 \ud2b9\uc131\uc740 \uac00\uc9c0\uace0 \uc788\uc9c0 \uc54a\ub2e4\uace0 \uc8fc\uc7a5\ud569\ub2c8\ub2e4. \ub530\ub77c\uc11c AI\ub97c \ub3c4\uad6c\ub85c \uc778\uc2dd\ud558\uace0, \uadf8 \ud55c\uacc4\ub97c \uc774\ud574\ud558\uba70 \uc778\uac04\uc774 \uc8fc\uccb4\uc801\uc73c\ub85c \ud65c\uc6a9\ud574\uc57c \ud55c\ub2e4\ub294 \uac83\uc774 \ud544\uc790\uc758 \uc8fc\uc7a5\uc785\ub2c8\ub2e4.\n1) AI\ub294 \uc778\uac04\uc758 \uc0dd\ud65c\uc744 \ud48d\uc694\ub86d\uac8c \ud558\ub294 \ub9cc\ub2a5 \ub3c4\uad6c\uc774\ub2e4: \ubcf8\ubb38\uc5d0\uc11c\ub294 AI\uac00 \ub9cc\ub2a5\uc774 \uc544\ub2c8\ub77c\uace0 \uba85\uc2dc\ud558\uace0 \uc788\uae30 \ub54c\ubb38\uc5d0 \ud2c0\ub9b0 \ub2f5\uc785\ub2c8\ub2e4.\n2) AI\ub294 \uc778\uac04\uc758 \ucc3d\uc758\uc131\uc774\ub098 \uc724\ub9ac\uad00\uc744 \ub2a5\uac00\ud558\ub294 \uc874\uc7ac\uac00 \ub420 \uac83\uc774\ub2e4: \ubcf8\ubb38\uc5d0\uc11c\ub294 AI\uac00 \uc778\uac04\uc758 \ucc3d\uc758\uc131, \uc724\ub9ac\uad00, \uac10\uc815 \ub4f1\uc744 \uac00\uc9c0\uace0 \uc788\uc9c0 \uc54a\ub2e4\uace0 \ud588\uc73c\ubbc0\ub85c \ud2c0\ub9b0 \ub2f5\uc785\ub2c8\ub2e4.\n3) AI\uc758 \ud65c\uc6a9\uc740 \uc0ac\ud68c\uc5d0 \uc608\uc0c1\uce58 \ubabb\ud55c \ud3d0\ud574\ub97c \uac00\uc838\uc624\uae30 \ub54c\ubb38\uc5d0 \uc2e0\uc911\ud574\uc57c \ud55c\ub2e4: AI \ud65c\uc6a9\uc758 \uc704\ud5d8\uc131\ub9cc \uac15\uc870\ud558\ub294 \uc120\ud0dd\uc9c0\uc774\uba70, \ubcf8\ubb38\uc5d0\uc11c\ub294 AI\ub97c \uc778\uac04\uc774 \uc8fc\uccb4\uc801\uc73c\ub85c \ud65c\uc6a9\ud574\uc57c \ud55c\ub2e4\ub294 \uc810\uc744 \uac15\uc870\ud558\uace0 \uc788\uc73c\ubbc0\ub85c \ud2c0\ub9b0 \ub2f5\uc785\ub2c8\ub2e4.",
-  "level": "N2",
-  "problem_type": "R"
-}
-백엔드 API로 문제 전송: http://localhost:8080/api/problems
-백엔드 전송 오류 발생: 500 Server Error:  for url: http://localhost:8080/api/problems
-INFO:     127.0.0.1:62934 - "POST /generate-problem HTTP/1.1" 500 Internal Server Error
